espresso10: build KEYBOARD_SEC_DOCK as module
Our sec_keyboard driver does not properly unregister the input device
when no dock is detected, which makes the system believe that a
physical keyboard is present at all times, which prevents
the software keyboard from being shown.

Some user own a Keyboard-Dock and meight miss the ability to use
the dock if it gets disabled completely.

Building it as module gives the user the ability to add a simple
init.d script to start the module at boot if they need it.

Change-Id: I12ca186efc43bdd84e10ea7d9cfef100c373e38b